[TF-239] Update java version to 19 and gradle to 8.1
