+++ .github/workflows/gradle.yml
@@ -3,9 +3,9 @@ name: Gradle Package
 on:
   push:
     branches:
-      - '**'
+      - "**"
     tags:
-      - '**'
+      - "**"
   release:
     types: [created]
 
@@ -17,27 +17,27 @@ jobs:
       packages: write
 
     steps:
-    - uses: actions/checkout@v2
-    - name: Set up JDK 8
-      uses: actions/setup-java@v2
-      with:
-        java-version: '8'
-        distribution: 'adopt'
-        server-id: github
-        settings-path: ${{ github.workspace }}
+      - uses: actions/checkout@v2
+      - name: Set up JDK 19
+        uses: actions/setup-java@v2
+        with:
+          java-version: "19"
+          distribution: "adopt"
+          server-id: github
+          settings-path: ${{ github.workspace }}
 
-    - name: Display bradle version
-      run: ./gradlew versionDisplay
+      - name: Display bradle version
+        run: ./gradlew versionDisplay
 
-    - name: Build with Gradle
-      run: ./gradlew build
+      - name: Build with Gradle
+        run: ./gradlew build
 
-    - name: Gradle report
-      run: ./gradlew jacocoTestReport coveralls
+      - name: Gradle report
+        run: ./gradlew jacocoTestReport coveralls
 
-    - name: Publish to GitHub Packages
-      if: startsWith(github.ref, 'refs/tags/')
-      run: ./gradlew publish
-      env:
-        GITHUB_USERNAME: ${{ github.actor }}
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      - name: Publish to GitHub Packages
+        if: startsWith(github.ref, 'refs/tags/')
+        run: ./gradlew publish
+        env:
+          GITHUB_USERNAME: ${{ github.actor }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
