+++ .tool-versions
@@ -1,0 +1,1 @@
+java adoptopenjdk-19.0.2+7
