+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).  
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).  
 
+## Unrelased
+
+* [TF-239] Update java version to 19 and gradle version to 8.1
+
 ## 1.8.0
 
 * [TT-11417] Implemented adjustable margin sizes in TicketPageSettings
