+++ build.gradle
@@ -2,23 +2,23 @@ plugins {
     id "com.jfrog.bintray" version "1.8.5"
     id "maven-publish"
     id 'java-library'
-    id 'net.nemerosa.versioning' version '2.7.1'
+    id 'net.nemerosa.versioning' version '3.0.0'
     id 'jacoco'
     id 'checkstyle'
-    id 'com.github.kt3k.coveralls' version '2.8.2'
+    id 'com.github.kt3k.coveralls' version '2.12.2'
 }
 group = 'au.com.sealink'
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = JavaVersion.VERSION_19
+targetCompatibility = JavaVersion.VERSION_19
 tasks.withType(JavaCompile) {
-	options.encoding = 'UTF-8'
+    options.encoding = 'UTF-8'
 }
 
 jacocoTestReport {
     reports {
-        xml.enabled = true // coveralls plugin depends on xml format report
-        html.enabled = true
+        xml.required.set(true) // coveralls plugin depends on xml format report
+        html.required.set(true)
     }
 }
 
@@ -44,6 +44,12 @@ checkstyle {
     toolVersion = '8.1'
 }
 
+checkstyle {
+    configProperties = [
+            'allowStaticMemberImports': 'true'
+    ]
+}
+
 repositories {
     google()
     mavenCentral()
@@ -51,13 +57,19 @@ repositories {
 }
 
 dependencies {
-    compile group: 'commons-codec', name: 'commons-codec', version:'1.6'
-    testCompile 'junit:junit:4.13.2'
+    implementation group: 'commons-codec', name: 'commons-codec', version:'1.15'
+    testImplementation 'junit:junit:4.13.2'
 }
 
-task sourcesJar(type: Jar, dependsOn: classes) {
-    classifier = 'sources'
+task sourcesJar(type: Jar) {
+    archiveClassifier.set('sources')
     from sourceSets.main.allSource
+}
+
+task versionInfo {
+    doLast {
+        println "Version: ${project.version}"
+    }
 }
 
 artifacts {
@@ -89,9 +101,7 @@ publishing {
     publications {
         mavenPublication(MavenPublication) {
             from components.java
-            artifact sourcesJar {
-                classifier "sources"
-            }
+            artifact sourcesJar
             groupId 'au.com.sealink'
             artifactId 'printing'
             version project.version
