+++ src/main/java/au/com/sealink/printing/ticketprinter/FontStyle.java
@@ -1,4 +1,4 @@
-package au.com.sealink.printing.ticket_printer;
+package au.com.sealink.printing.ticketprinter;
 
 public enum FontStyle {
     Bold, Italic, Inverted
