+++ src/main/java/au/com/sealink/printing/ticketprinter/TicketPrinter.java
@@ -1,14 +1,14 @@
-package au.com.sealink.printing.ticket_printer;
+package au.com.sealink.printing.ticketprinter;
 
+import au.com.sealink.printing.ticketprinter.exceptions.NoSuchPrinterException;
+import au.com.sealink.printing.ticketprinter.exceptions.NoTicketPageSettingsAssigned;
+import au.com.sealink.printing.ticketprinter.printables.PrintableTickets;
 import java.awt.print.Printable;
 import java.awt.print.PrinterException;
 import java.awt.print.PrinterJob;
 import java.util.List;
 import javax.print.PrintService;
 
-import au.com.sealink.printing.ticket_printer.exceptions.NoSuchPrinterException;
-import au.com.sealink.printing.ticket_printer.exceptions.NoTicketPageSettingsAssigned;
-import au.com.sealink.printing.ticket_printer.printables.PrintableTickets;
 
 /*
  * TicketPrintCommand
