+++ src/main/java/au/com/sealink/printing/ticketprinter/Underline.java
@@ -1,4 +1,4 @@
-package au.com.sealink.printing.ticket_printer;
+package au.com.sealink.printing.ticketprinter;
 
 public enum Underline {
     None,
