+++ src/main/java/au/com/sealink/printing/ticketprinter/exceptions/NoTicketPageSettingsAssigned.java
@@ -1,4 +1,4 @@
-package au.com.sealink.printing.ticket_printer.exceptions;
+package au.com.sealink.printing.ticketprinter.exceptions;
 
 public class NoTicketPageSettingsAssigned extends Exception {
