+++ src/main/java/au/com/sealink/printing/ticketprinter/printables/PrintableTextElement.java
@@ -1,9 +1,8 @@
-package au.com.sealink.printing.ticket_printer.printables;
+package au.com.sealink.printing.ticketprinter.printables;
 
+import au.com.sealink.printing.ticketprinter.TicketElement;
 import java.awt.Font;
 import java.awt.Graphics2D;
-
-import au.com.sealink.printing.ticket_printer.TicketElement;
 
 class PrintableTextElement extends PrintableElement {
