+++ src/test/java/au/com/sealink/printing/utils/NumberConverterTest.java
@@ -1,13 +1,13 @@
 package au.com.sealink.printing.utils;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 
 public class NumberConverterTest {
 
-  @Test
-  public void testMm2finch() {
-    assertEquals(NumberConverter.mm2finch(10), 28.34, 0.01);
-  }
+    @Test
+    public void testMm2finch() {
+        assertEquals(NumberConverter.mm2finch(10), 28.34, 0.01);
+    }
 }
